Common methods for all shc devices are moved to entity class. Moved .translations to translations folder.
