+++ homeassistant/components/boschshc/binary_sensor.py
@@ -14,6 +14,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_platform
 
 from .const import DOMAIN
+from .entity import SHCEntity
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -58,79 +59,8 @@ async def async_setup_entry(hass, config_entry, async_add_entities):
     )
 
 
-class ShutterContactSensor(BinarySensorDevice):
+class ShutterContactSensor(SHCEntity, BinarySensorDevice):
     """Representation of a SHC shutter contact sensor."""
-
-    def __init__(self, device: SHCShutterContact, room_name: str, controller_ip: str):
-        """Initialize the SHC device."""
-        self._device = device
-        self._room_name = room_name
-        self._controller_ip = controller_ip
-
-    async def async_added_to_hass(self):
-        """Subscribe to SHC events."""
-        await super().async_added_to_hass()
-
-        def on_state_changed():
-            self.schedule_update_ha_state()
-
-        for service in self._device.device_services:
-            service.subscribe_callback(self.entity_id, on_state_changed)
-
-    async def async_will_remove_from_hass(self):
-        """Unsubscribe from SHC events."""
-        await super().async_will_remove_from_hass()
-        for service in self._device.device_services:
-            service.unsubscribe_callback(self.entity_id)
-
-    @property
-    def unique_id(self):
-        """Return the unique ID of this binary sensor."""
-        return self._device.serial
-
-    @property
-    def device_id(self):
-        """Return the ID of this binary sensor."""
-        return self._device.id
-
-    @property
-    def root_device(self):
-        """Return the root device id."""
-        return self._device.root_device_id
-
-    @property
-    def name(self):
-        """Name of the device."""
-        return self._device.name
-
-    @property
-    def manufacturer(self):
-        """Manufacturer of the device."""
-        return self._device.manufacturer
-
-    @property
-    def device_info(self):
-        """Return the device info."""
-        return {
-            "identifiers": {(DOMAIN, self.device_id)},
-            "name": self.name,
-            "manufacturer": self.manufacturer,
-            "model": self._device.device_model,
-            "sw_version": "",
-            "via_device": (DOMAIN, self._controller_ip),
-        }
-
-    @property
-    def should_poll(self):
-        """Set polling mode."""
-        return False
-
-    @property
-    def available(self):
-        """Return false if status is unavailable."""
-        if self._device.status == "AVAILABLE":
-            return True
-        return False
 
     @property
     def is_on(self):
@@ -152,21 +82,8 @@ class ShutterContactSensor(BinarySensorDevice):
         }
         return switcher.get(self._device.device_class, DEVICE_CLASS_WINDOW)
 
-    def update(self):
-        """Trigger an update of the device."""
-        self._device.update()
 
-    @property
-    def state_attributes(self):
-        """Extend state attribute of the device."""
-        state_attr = super().state_attributes
-        if state_attr is None:
-            state_attr = dict()
-        state_attr["boschshc_room_name"] = self._room_name
-        return state_attr
-
-
-class SmokeDetectorSensor(BinarySensorDevice):
+class SmokeDetectorSensor(SHCEntity, BinarySensorDevice):
     """Representation of a SHC smoke detector sensor."""
 
     def __init__(
@@ -177,76 +94,10 @@ class SmokeDetectorSensor(BinarySensorDevice):
         hass: HomeAssistant,
     ):
         """Initialize the SHC device."""
-        self._device = device
-        self._room_name = room_name
-        self._controller_ip = controller_ip
+        super().__init__(
+            device=device, room_name=room_name, controller_ip=controller_ip
+        )
         self._hass = hass
-
-    async def async_added_to_hass(self):
-        """Subscribe to SHC events."""
-        await super().async_added_to_hass()
-
-        def on_state_changed():
-            self.schedule_update_ha_state()
-
-        for service in self._device.device_services:
-            service.subscribe_callback(self.entity_id, on_state_changed)
-
-    async def async_will_remove_from_hass(self):
-        """Unsubscribe from SHC events."""
-        await super().async_will_remove_from_hass()
-        for service in self._device.device_services:
-            service.unsubscribe_callback(self.entity_id)
-
-    @property
-    def unique_id(self):
-        """Return the unique ID of this binary sensor."""
-        return self._device.serial
-
-    @property
-    def device_id(self):
-        """Return the ID of this binary sensor."""
-        return self._device.id
-
-    @property
-    def root_device(self):
-        """Return the root device id."""
-        return self._device.root_device_id
-
-    @property
-    def name(self):
-        """Name of the device."""
-        return self._device.name
-
-    @property
-    def manufacturer(self):
-        """Manufacturer of the device."""
-        return self._device.manufacturer
-
-    @property
-    def device_info(self):
-        """Return the device info."""
-        return {
-            "identifiers": {(DOMAIN, self.device_id)},
-            "name": self.name,
-            "manufacturer": self.manufacturer,
-            "model": self._device.device_model,
-            "sw_version": "",
-            "via_device": (DOMAIN, self._controller_ip),
-        }
-
-    @property
-    def should_poll(self):
-        """Report polling mode."""
-        return False
-
-    @property
-    def available(self):
-        """Return false if status is unavailable."""
-        if self._device.status == "AVAILABLE":
-            return True
-
-        return False
 
     @property
     def is_on(self):
@@ -261,17 +112,12 @@ class SmokeDetectorSensor(BinarySensorDevice):
         """Return the class of this device, from component DEVICE_CLASSES."""
         return DEVICE_CLASS_SMOKE
 
-    def update(self):
-        """Trigger an update of the device."""
-        self._device.update()
-
     @property
     def state_attributes(self):
         """Extend state attribute of the device."""
         state_attr = super().state_attributes
         if state_attr is None:
             state_attr = dict()
-        state_attr["boschshc_room_name"] = self._room_name
         state_attr["boschshc_smokedetector_checkstate"] = (
             "OK"
             if self._device.smokedetectorcheck_state
