+++ homeassistant/components/boschshc/translations/de.json
