+++ homeassistant/components/boschshc/translations/en.json
